code/lambdas/export-lambda/requirements.txt
boto3==1.34.69
python-docx==1.1.0
python-pptx==0.6.23
reportlab==4.1.0
requests==2.32.3
beautifulsoup4==4.13.4
pillow==11.2.1
plotly
numpy==2.2.0
matplotlib